% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_save.R
\name{gf_save}
\alias{gf_save}
\alias{gf_load}
\title{Save and load encrypted files}
\usage{
gf_save(
  ...,
  list = character(),
  file = stop("file must be specified (.rdg file extension is recommended)"),
  user = character(0),
  local_user = TRUE,
  ascii = FALSE,
  version = NULL,
  compress = "xz",
  overwrite = FALSE
)

gf_load(
  file = stop("file must be specified (.rdg file extension is recommended)")
)
}
\description{
Save and load encrypted files
}
