% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_saveRDS.R
\name{gf_saveRDS}
\alias{gf_saveRDS}
\alias{gf_readRDS}
\title{Save and read encrypted RDS}
\usage{
gf_saveRDS(
  object,
  file = stop("file must be specified (.rdg file extension is recommended)"),
  user = character(0),
  local_user = TRUE,
  comment = "",
  ascii = FALSE,
  version = NULL,
  compress = "xz",
  overwrite = FALSE
)

gf_readRDS(
  file = stop("file must be specified (.rdg file extension is recommended)")
)
}
\arguments{
\item{overwrite}{}
}
\description{
Save and read encrypted RDS
}
