% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_check.R
\name{gy_check}
\alias{gy_check}
\alias{gy_userfile}
\title{Utilities for goldfinger encryption}
\usage{
gy_userfile(path = getOption("goldeneye_path"), silent = FALSE)

gy_check(local = NULL, silent = FALSE)
}
\description{
Utilities for goldfinger encryption
}
