% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_resave.R
\name{gy_resave}
\alias{gy_resave}
\title{Upgrade older goldfinger saved data files to remove dependencies on the qs package}
\usage{
gy_resave(
  input_file = stop("input_file must be specified (.rdg file extension is recommended)"),
  output_file = input_file
)
}
\arguments{
\item{input_file}{the name of the saved file to resave}

\item{output_file}{the name of the file to save to (default is to replace the existing file)}
}
\description{
Upgrade older goldfinger saved data files to remove dependencies on the qs package
}
\details{
For more details see vignette("goldfinger-qs")
}
