% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_save.R
\name{gy_save}
\alias{gy_save}
\alias{gy_load}
\title{Save and load encrypted files}
\usage{
gy_save(
  ...,
  list = character(),
  file = stop("file must be specified (.rdg file extension is recommended)"),
  user = character(0),
  local_user = TRUE,
  comment = "",
  overwrite = FALSE,
  ascii = FALSE,
  funs = list(type = "identity"),
  method = "qs"
)

gy_load(
  file = stop("file must be specified (.rdg file extension is recommended)")
)
}
\description{
Save and load encrypted files
}
