% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_saveRDS.R
\name{gf_saveRDS}
\alias{gf_saveRDS}
\alias{gy_saveRDS}
\alias{gy_readRDS}
\title{Save and read encrypted RDS}
\usage{
gy_saveRDS(
  object,
  file = stop("file must be specified (.rdg file extension is recommended)"),
  user = character(0),
  local_user = TRUE,
  comment = "",
  overwrite = FALSE,
  ascii = FALSE,
  funs = list(type = "identity"),
  ...
)

gy_readRDS(
  file = stop("file must be specified (.rdg file extension is recommended)"),
  run_custom = TRUE
)
}
\description{
Save and read encrypted RDS
}
