% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_setup.R
\name{gy_setup}
\alias{gy_setup}
\title{Title}
\usage{
gy_setup(
  weblink = NULL,
  name = NULL,
  email = NULL,
  user = NULL,
  filename = NULL,
  path = NULL,
  append_Rprofile = TRUE
)
}
\description{
Title
}
