% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_save.R
\name{gf_save}
\alias{gf_save}
\title{Title}
\usage{
gf_save(
  ...,
  list = character(),
  file = "encrypted_data.rdg",
  user = gf_user(),
  compress = "xz"
)
}
\arguments{
\item{user}{}
}
\description{
Title
}
