% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_check.R
\name{gf_utilities}
\alias{gf_utilities}
\alias{gf_check}
\alias{gf_user}
\title{Utilities for goldfinge encryption}
\usage{
gf_check(silent = FALSE)

gf_user()
}
\arguments{
\item{silent}{}
}
\description{
Utilities for goldfinge encryption
}
