% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_utilities.R
\name{gf_utilities}
\alias{gf_utilities}
\alias{gf_check}
\title{Utilities for goldfinger encryption}
\usage{
gf_check(path = getOption("goldfinger_path"), silent = FALSE)
}
\arguments{
\item{silent}{}
}
\description{
Utilities for goldfinger encryption
}
