% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldfinger-package.R
\docType{package}
\name{goldfinger-package}
\alias{goldfinger}
\alias{goldfinger-package}
\title{goldfinger: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
