% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldfinger-package.R
\docType{package}
\name{goldfinger-package}
\alias{goldfinger}
\alias{goldfinger-package}
\title{goldfinger: Support Package for GOLDFINGER}
\description{
Support package for GOLDFINGER
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Mossa Merhi Reimert \email{mossa@sund.ku.dk} [contributor]
  \item Søren Saxmose Nielsen \email{saxmose@sund.ku.dk} [contributor]
}

}
\keyword{internal}
