% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_check.R
\name{gy_check}
\alias{gy_check}
\title{Utilities for goldfinger encryption}
\usage{
gy_check(path = getOption("goldfinger_path"), silent = FALSE)
}
\arguments{
\item{silent}{}
}
\description{
Utilities for goldfinger encryption
}
