% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_encrypt.R
\name{gy_encrypt}
\alias{gy_encrypt}
\title{Encrypt and decrypt a pre-serialised object using goldeneye}
\usage{
gy_encrypt(
  object,
  user = character(0),
  local_user = TRUE,
  comment = "",
  funs = list(type = "identity")
)
}
\arguments{
\item{overwrite}{}
}
\description{
Encrypt and decrypt a pre-serialised object using goldeneye
}
